refactor: move `spdx#Satisfies` into internal package
