+++ internal/spdx/satisfies.go
