+++ pkg/osvscanner/vulnerability_result.go
@@ -8,10 +8,10 @@ import (
 	"github.com/google/osv-scanner/internal/grouper"
 	"github.com/google/osv-scanner/internal/output"
 	"github.com/google/osv-scanner/internal/sourceanalysis"
+	"github.com/google/osv-scanner/internal/spdx"
 	"github.com/google/osv-scanner/pkg/models"
 	"github.com/google/osv-scanner/pkg/osv"
 	"github.com/google/osv-scanner/pkg/reporter"
-	"github.com/google/osv-scanner/pkg/spdx"
 )
 
 // buildVulnerabilityResults takes the responses from the OSV API and the deps.dev API
